Corrected formatting for planetary received flux calculation

+++ hab_zone.py
@@ -29,7 +29,7 @@ class HabZoneEvaluator:
         for index, row in data.iterrows():
             semimajor = explorer._orb_dist(row)
             t_star = row['st_teff'] - 5780
-            pl_stflux = (10**row['st_lum'])/semimajor**2 #Stellar luminosity is in units of log(L/L_sun)
+            pl_stflux = float((10**row['st_lum'])/semimajor**2) #Stellar luminosity is in units of log(L/L_sun)
             cons_inner_stflux = (sol_flux_in + a_in*t_star + b_in*(t_star**2) + c_in*(t_star**3) + d_in*(t_star**4))/np.sqrt(1 - row['pl_orbeccen']**2)
             cons_outer_stflux = (sol_flux_out + a_out*t_star + b_out*(t_star**2) + c_out*(t_star**3) + d_out*(t_star**4))/np.sqrt(1 - row['pl_orbeccen']**2)
             cons_inner_rad = np.sqrt((10**row['st_lum'])/cons_inner_stflux)
@@ -64,7 +64,7 @@ class HabZoneEvaluator:
         for index, row in data.iterrows():
             semimajor = explorer._orb_dist(row)
             t_star = row['st_teff'] - 5780
-            pl_stflux = (10**row['st_lum'])/semimajor**2 #Stellar luminosity is in units of log(L/L_sun)
+            pl_stflux = float((10**row['st_lum'])/semimajor**2) #Stellar luminosity is in units of log(L/L_sun)
             opt_inner_stflux = (sol_flux_in + a_in*t_star + b_in*(t_star**2) + c_in*(t_star**3) + d_in*(t_star**4))/np.sqrt(1 - row['pl_orbeccen']**2)
             opt_outer_stflux = (sol_flux_out + a_out*t_star + b_out*(t_star**2) + c_out*(t_star**3) + d_out*(t_star**4))/np.sqrt(1 - row['pl_orbeccen']**2)
             opt_inner_rad = np.sqrt((10**row['st_lum'])/opt_inner_stflux)
